requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
scikit-image==0.21.0
mediapipe==0.10.21
Pillow==10.0.0
tqdm==4.65.0
pyyaml==6.0.1
seaborn==0.12.2
albumentations==1.3.1